Create the needed folders to run GMACS

Function to create empty folders needed to run GMACS: /build, /build/debug, /build/release, and a folder for the species (Spc) of interest (/build/Spc)

SetGMACSfold(vv = NULL, Dir = NULL, Spc = NULL, Sp = NULL, verbose = FALSE)

Arguments

vv

integer index in Dir indicating root folder to create these base folders.

Dir

vector of root folder names

Spc

vector of species names

Sp

integer index in Spc indicating the name(s) of the species for which the folder needs to be created. If Sp = "all", then a folder will be created for all species designated in Spc.

verbose

flag (TRUE/FALSE) to print processing information

Value

nothing

Details

Both the input and output files of GMACS will be contained in the folder whose name corresponds to the species of interest. When starting from scratch, you will need to provide this folder with all the assessment files needed by GMACS (.dat, .ctl, and .prj files).