.buildGMACS

Function to build the GMACS executable. It converts the gmacs.tpl code to a gmacs.cpp code, calls ADMB to compile all .cpp files required to build gmacs and returns the gmacs executable.

.buildGMACS(
  prefix = NULL,
  raneff = NULL,
  safe = NULL,
  dll = NULL,
  debug = NULL,
  logfile = NULL,
  add = NULL,
  verbose = NULL,
  pathfile = NULL,
  args = NULL
)

Arguments

prefix

String name prefix (i.e., without extension) of the ADMB project; here "gmacs".

raneff

logical: if TRUE, use the random effects model executable tpl2rem.exe, otherwise use the normal model executable tpl2cpp.exe.

safe

logical: if TRUE, use safe mode with bounds checking on all array objects, otherwise use optimized mode for fastest execution.

dll

create dll (rather than executable)

debug

compile with debug symbols

logfile

logical: if TRUE, create a log file of the messages from the shell call.

add

logical: if TRUE, append shell call messages to an exsiting log file.

verbose

logical: if TRUE, report the shell call an its messages to the R console.

pathfile

string name of 2-column text file that details the relevant paths for the R variables admbpath, gccpath, and editor.

args

list of string names corresponding to the various .cpp libraries required to build gmacs. These files are stored in the lib/ folder.

Value

nothing.