Read all input and output files

This function reads all the Gmacs input files ("gmacs.dat", "model.dat", "model.ctl", "model.prj" and if they exist, all the output files ("Gmacsall.out", "gmacs.par", "gmacs.rep", "simdata.out")).

getInpOutFiles(Dir = NULL, verbose = NULL)

Arguments

Dir

(character string)- path where the input/out files are saved for the stock of interest.

Value

a named list of lists:

GMACSdat

the gmacs.dat file;

datFile

the model.dat file;

ctlFile

the model.ctl file;

prjfile

the model.prj file;

GmacsAll_out

the Gmacsall.out file if available;

gmacs_par

the gmacs.par file if available;

gmacs_rep

the gmacs.rep file if available;

gmacs_sim

the simdata.out file if available

See also

Examples

if (FALSE) {
# Load package ----
library(gmr)

# Set directories ----
Dir_Dvpt_Vers <- file.path(here::here(), "Dvpt_Version", fsep = fsep)
Dir_Last_Vers <- file.path(here::here(), "Latest_Version", fsep = fsep)

# Define the stock of interest ----
stock <- "SMBKC"

# Read Gmacs files ----
Stock_files <-
getInpOutFiles(Dir = file.path(Dir_Dvpt_Vers, 'build', stock, fsep = fsep),
verbose = TRUE)
}