Read the GMACS data file

Read the GMACS data file.

readGMACSdat(FileName = NULL, verbose = TRUE)

Arguments

FileName

(character string)- path (and name) of the data file (e.g. snow.data)

verbose

(logical)- (TRUE/FALSE); flag to print processing information

Value

A list with the content of the model.dat file.

sourcefile - The file source.

Comments - Specifications about the GMACS version used and the stock assessed.

Start_Y - first year in the model.

End_Y - Last year in the model.

N_year - Number of years in total (i.e., End_Y-DatOut+ 1).

N_seasons - Number of seasons in the model.

N_fleet - Number of fleet(s) in the model.

N_sexes - Number of sex(es) in the model.

N_shell_cdt - Number of shell condition in the model.

N_maturity - Number of maturity state(s) in the model.

N_sizeC - Number of size class in the model.

Recr_Season - The season recruitment occurs.

Grwth_Season - The season growth occurs.

SSB_Season - The season to calculate the Spawning Stock Biomass.

N_Season - The seaspn for the number (N) output.

Max_sizeC - The maximum size class in the model.

Size_breaks - The size breaks considered in the population.

M_in_Type - The natural mortality per season input type (vector or matrix).

M_Seas_prop - The proportion of natural mortality per season.

F_Fleet_names - The names of the fishing fleets.

Survey_names - The names of the surveys.

F_Season_Type - The type of fishing mortality per season (0: instantaneous; 1: continuous).

CatchDF_format - The input format for catch data (0: old format; 1: new format).

N_CatchDF - The number of catch data frame.

Nrows_CatchDF - The number of rows in each catch data frame.

Catch - The catch data.

SurveyDF_format - The input format for survey data (0: old format; 1: new format).

N_SurveyDF - The number of survey data frame.

Sv_type - The data type for each abundance index (1: total selectivity; 2:retention*selectivity).

Nrows_SvDF - The number of rows in each survey data frame.

Surveys - Teh survey data.

SizeFreqDF_format - The input format for size composition data (0: old format; 1: new format).

N_SizeFreq_df - The number of size composition data frame.

Nrows_SiseFreqDF - The number of rows in each size composition data frame.

Nbins_SiseFreq - The number of bins in each size composition data frame.

SizeFreq - The size composition data.

GrowthObsType - The type of observation for growth (increment or change in size-class)

NGrowthObs - The number of observations (lines) for each growth data frame.

GrowthData - The growth data.

NenvIndics - The number of environmental indices considered.

EnvYrs - The years cover for each environmental index.

EnvData - Teh environmental data.

eof - Logical indicating the end of the file (used for checking the reading).

See also

Examples

if (FALSE) {
# Stock ----
stock <- "SNOW_crab"
# GMACS input files ----
datfileName <- "snow_21_M09.dat"
ctlfileName <- "snow_21_M09.ctl"
# read gmacs.dat ----
fileName <- "gmacs.dat"
fileName <- file.path(dir_Base, stock, fileName, fsep = fsep)
GMACSdat <- readGMACS.dat(path = fileName, verbose = TRUE)
# Read the data file ----
datFile <- file.path(dir_Base, stock, datfileName, fsep = fsep)
datFile <- readGMACSdat(FileName = datFile, verbose = T)
}