readGMACSprj

Read the GMACS projection file.

readGMACSprj(FileName = NULL, verbose = TRUE)

Arguments

FileName

path (and name) of the projection file (e.g. snow.prj)

verbose

(TRUE/FALSE); flag to print processing information

Value

A list with the content of the model.prj file.

sourcefile - The file source.

Comments - Specifications about the GMACS version used and the stock assessed.

Calc_MSY - Compute MSY ?

Ffixed - The mortality rate applied to each fishery.

spr_syr - The first year for average recruitment/MMB for Bspr calculation.

spr_nyr - The last year for average recruitment/MMB for Bspr calculation.

spr_SexR_syr - The first year for computing the sex ratio used in the calculation of the BRPs.

spr_SexR_nyr - The last year for computing the sex ratio used in the calculation of the BRPs.

spr_aveF_syr - The first year for computing the average fishing mortality for discards.

spr_aveF_nyr - The last year for computing the average fishing mortality for discards.

spr_M_syr - The first year for computing the natural mortality.

spr_M_nyr - The last year for computing the natural mortality.

spr_Prop_syr - The first year for computing the season length.

spr_Prop_nyr - The last year for computing the season lengths.

spr_grow_yr - The first year for computing growth.

spr_sel_syr - The first year for computing the average vulnearbility.

spr_sel_nyr - The last year for computing the average vulnearbility.

spr_target - The target for teh SPR ratio for Bmsy proxy.

OFLTier - The tier system considered.

OFLalpha - The alpha value used for that tier system (i.e., teh cut-off).

OFLbeta - The beta value used for that tier system (i.e., the limit).

OFLgamma - The gamma value used in the tier system.

ABCBuffer - The ABC-OFL buffer to consider.

Compute_yield_prj - Is the yield function reported ?

pyr - The last year of the projection period.

prj_type - The projection type (Constant F; proportion of current F).

prj_Nstrat - The number of strategies considered in the projection.

prj_Frange - The range of F values for the strategies considered in the projections.

prj_bycatch_on - Allow for bycatch fleets to have non-zero mortality in the projections.

prj_replicates - How many times each MCM draw is run in the projections.

Fixed_prj_Bmsy - Should Bmsy be fixed?

proj_syr - The first year for computing the average recruitment in the projections.

proj_nyr - The last year for computing the average recruitment in the projections.

proj_SexR_syr - The first year for computing the average sex-ratio in the projections.

proj_SexR_nyr - The last year for computing the average sex-ratio in the projections.

proj_aveF_syr - The first year for computing the average fishing moratlity for discards in the projections.

proj_aveF_nyr - The last year for computing the average fishing moratlity for discards in the projections.

proj_M_syr - The first year for computing the natural mortality in the projections.

proj_M_nyr - The last year for computing the natural mortality in the projections.

proj_Prop_syr - The first year for computing the season lengths in the projections.

proj_Prop_nyr - The last year for computing the season lengths in the projections.

proj_grow_yr - The year for specifying growth in the projections.

proj_sel_syr - The first year for computing the average vulnerability in the projections.

proj_sel_nyr - The last year for computing the average vulnerability in the projections.

Stock_rec_prj - The stock recruitment option for the projections.

Age_at_rec_prj - The age at recruitment in the projections.

prj_futRec_syr - The first year for generating recruitment in the projections.

prj_futRec_nyr - The last year for generating recruitment in the projections.

mean_rec_prj - The mean recruitment for the projections.

SigmaR_prj - The sigma used for recruitment in the projections.

Prow_prj - A scalor in for the recruitment deviations used in the projections.

Initial_eps - The first recruitment deviations.

Apply_HCR_prj - Which strategy to apply? (Harvest Control Rules strategies; constant F)

Apply_StateStrat_prj - Apply the state strategy?

Nb_state_param - The number of state strategy parameters.

MeanWStateMature - The mean weight of the mature individuals.

MeanWStateLegal - The mean (legal) weight of the mature individuals.

max_prj - The number of function call (stop after max_prj mcdraws).

full_prj_diag - Output a full diagnostic?

eof - Logical indicating the end of the file (used for checking the reading).

See also

Examples

if (FALSE) {
# Stock ----
stock <- "SNOW_crab"
# GMACS input files ----
prjfileName <- "snow_21_M09.prj"
# Read the projection file ----
prjfile <- file.path(dir_Base, stock, prjfileName, fsep = fsep)
prjfile <- readGMACSprj(FileName = prjfile, verbose = T)
}